Add an observation in trait measurement table

Add a trait measure in trait measurement table

Usage

add_traits_measures(
  new_data,
  col_names_select = NULL,
  col_names_corresp = NULL,
  collector_field = NULL,
  plot_name_field = NULL,
  individual_plot_field = NULL,
  id_plot_name = NULL,
  id_tag_plot = NULL,
  id_specimen = NULL,
  traits_field,
  features_field = NULL,
  allow_multiple_value = FALSE,
  add_data = FALSE
)

Arguments

new_data

tibble

col_names_select

string vector

col_names_corresp

string vector

collector_field

string column name which contain the collector name

plot_name_field

string column name which contain the plot_name for linking

individual_plot_field

string column name which contain the individual tag for linking

id_plot_name

string column name which contain the ID of plot_name

id_tag_plot

string column name which contain the ID of individuals table

id_specimen

string column name which contain the ID of specimen

traits_field

string vector listing trait columns names in new_data

features_field

string vector listing features (column names) to link to measurementsin new_data

allow_multiple_value

if multiple values linked to one individual can be uploaded at once

add_data

logical whether or not data should be added - by default FALSE

Value

list of tibbles that should be/have been added

Author

Gilles Dauby, gilles.dauby@ird.fr