Add link between specimen and individual

Generate link between individual and specimens

Usage

.add_link_specimens(
  new_data,
  col_names_select = NULL,
  col_names_corresp = c("id_specimen", "id_n", "type"),
  launch_adding_data = FALSE
)

Arguments

new_data

tibble fuzzy person name to look for

col_names_select

a vector of string that select columns of new_data to consider, it must be 3 columns

col_names_corresp

a vector of string of same length of col_names_select, should not be changed

launch_adding_data

logical, if TRUE links are added, by default it is FALSE for security

Value

A tibble of all subplots

Author

Gilles Dauby, gilles.dauby@ird.fr